Refactor struct rules into separate files
This PR reorganizes the struct rule implementation by moving extension-specific
code from _struct_spec.py into separate files:

- _struct_spec_dataclass.py: Dataclass-specific code and utilities
- _struct_spec_attrs.py: Attrs-related code
- _struct_spec_pydantic.py: Pydantic models support
- _struct_spec_flax.py: Flax-specific utilities

The main _struct_spec.py file now focuses on core registry functionality and
registers rules from the extension modules. This makes the codebase more
maintainable and easier to extend with new rules.
